image/requirements.txt
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
Flask==2.0.1
google-auth==2.0.1
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
kubernetes==18.20.0
MarkupSafe==2.0.1
oauthlib==3.1.1
psycopg2-binary==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.2
PyYAML==5.4.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.16.0
urllib3==1.26.6
websocket-client==1.2.1
Werkzeug==2.0.1
